+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "lib/core/msdf-sys",
     "lib/data",
     "lib/eval-tt",
+    "lib/logger",
     "lib/optics",
     "lib/prelude",
     "lib/code-builder",
