+++ lib/core/src/display/object.rs
@@ -9,9 +9,7 @@ use crate::closure;
 use crate::data::dirty;
 use crate::data::dirty::traits::*;
 use data::opt_vec::OptVec;
-use crate::system::web::group;
-
-use basegl_system_web::Logger;
+use logger::*;
 use nalgebra::Vector3;
 use nalgebra::Matrix4;
 use transform::CachedTransform;
