+++ lib/core/src/display/world/scene.rs
@@ -4,7 +4,7 @@ use crate::prelude::*;
 
 use crate::display::camera::Camera2D;
 use crate::display::object::DisplayObjectData;
-use basegl_system_web::Logger;
+use logger::*;
 
 
 // =============
