+++ lib/core/src/system/gpu/data/uniform.rs
@@ -9,7 +9,7 @@ use web_sys::WebGlUniformLocation;
 
 use crate::display::render::webgl::Context;
 use crate::system::gpu::data::texture::*;
-use crate::system::web::Logger;
+
