+++ lib/system/web/src/lib.rs
@@ -58,244 +58,6 @@ impl Error {
 }
 
 
-// ===================
-// === JS Bindings ===
-// ===================
-
-#[macro_export]
-macro_rules! console_log {
-    ($($t:tt)*) => ($crate::console::log_1(&format_args!($($t)*)
-                                    .to_string().into()))
-}
-
-
-// ==============
-// === LogMsg ===
-// ==============
-
-pub trait LogMsg {
-    fn with_log_msg<F: FnOnce(&str) -> T, T>(&self, f:F) -> T;
-}
-
-impl LogMsg for &str {
-    fn with_log_msg<F: FnOnce(&str) -> T, T>(&self, f:F) -> T {
-        f(self)
-    }
-}
-
-impl<F: Fn() -> S, S: AsRef<str>> LogMsg for F {
-    fn with_log_msg<G: FnOnce(&str) -> T, T>(&self, f:G) -> T {
-        f(self().as_ref())
-    }
-}
-
-
-// ==============
-// === Logger ===
-// ==============
-
-#[derive(Clone,Debug,Default)]
-pub struct Logger {
-    pub path: String,
-}
-
-#[allow(dead_code)]
-impl Logger {
-    pub fn new<T: AsRef<str>>(path:T) -> Self {
-        let path = path.as_ref().to_string();
-        Self { path }
-    }
-
-    pub fn sub<T: AsRef<str>>(&self, path: T) -> Self {
-        Self::new(format!("{}.{}", self.path, path.as_ref()))
-    }
-
-    pub fn group<M: LogMsg, T, F: FnOnce() -> T>(&self, msg: M, f: F) -> T {
-        self.group_begin(msg);
-        let out = f();
-        self.group_end();
-        out
-    }
-
-    fn format<M: LogMsg>(&self, msg: M) -> JsValue {
-        msg.with_log_msg(|s| format!("[{}] {}", self.path, s)).into()
-    }
-}
-
-#[cfg(target_arch = "wasm32")]
-impl Logger {
-    pub fn trace<M: LogMsg>(&self, _msg: M) {
-//        console::debug_1(&self.format(msg));
-    }
-
-    pub fn info<M: LogMsg>(&self, _msg: M) {
-//        console::group_1(&self.format(msg));
-//        console::group_end();
-    }
-
-    pub fn warning<M: LogMsg>(&self, msg: M) {
-        console::warn_1(&self.format(msg));
-    }
-
-    pub fn error<M: LogMsg>(&self, msg: M) {
-        console::error_1(&self.format(msg));
-    }
-
-    pub fn group_begin<M: LogMsg>(&self, _msg: M) {
-//        console::group_1(&self.format(msg));
-    }
-
-    pub fn group_end(&self) {
-//        console::group_end();
-    }
-}
-
-// FIXME: Add the non-wasm impl
-#[cfg(not(target_arch = "wasm32"))]
-impl Logger {
-    pub fn trace<M: LogMsg>(&self, _msg: M) {}
-    pub fn info<M: LogMsg>(&self, _msg: M) {}
-    pub fn warning<M: LogMsg>(&self, _msg: M) {}
-    pub fn error<M: LogMsg>(&self, _msg: M) {}
-    pub fn group_begin<M: LogMsg>(&self, _msg: M) {}
-    pub fn group_end(&self) {}
-}
-
-
-// ====================
-// === Logger Utils ===
-// ====================
-
-#[macro_export]
-macro_rules! fmt {
-    ($($arg:tt)*) => (||(format!($($arg)*)))
-}
-
-#[macro_export]
-macro_rules! group {
-    ($logger:expr, $message:expr, $body:tt) => {{
-        let __logger = $logger.clone();
-        __logger.group_begin(|| $message);
-        let out = $body;
-        __logger.group_end();
-        out
-    }};
-    ($logger:expr, $str:expr, $a1:expr, $body:tt) => {{
-        group!($logger, format!($str,$a1), $body)
-    }};
-    ($logger:expr, $str:expr, $a1:expr, $a2:expr, $body:tt) => {{
-        group!($logger, format!($str,$a1,$a2), $body)
-    }};
-    ($logger:expr, $str:expr, $a1:expr, $a2:expr, $a3:expr, $body:tt) => {{
-        group!($logger, format!($str,$a1,$a2,$a3), $body)
-    }};
-}
-
-//#[macro_export]
-//macro_rules! info {
-//    ($logger:expr, $message:expr) => {{
-//        $logger.info(||iformat!($message));
-//    }};
-//    ($logger:expr, $message:expr, $body:tt) => {{
-//        let __logger = $logger.clone();
-//        __logger.group_begin(||iformat!($message));
-//        let out = $body;
-//        __logger.group_end();
-//        out
-//    }};
-//}
-
-//#[macro_export]
-//macro_rules! warning {
-//    ($logger:expr, $message:expr) => {{
-//        $logger.warning(||iformat!($message));
-//    }};
-//    ($logger:expr, $message:expr, $body:tt) => {{
-//        let __logger = $logger.clone();
-//        __logger.group_begin(||iformat!($message));
-//        let out = $body;
-//        __logger.group_end();
-//        out
-//    }};
-//}
-
-
-
-
-
-
-
-
-#[macro_export]
-macro_rules! log_template {
-    ($method:ident $logger:expr, $message:tt $($rest:tt)*) => {
-        $crate::log_template_impl! {$method $logger, iformat!($message) $($rest)*}
-    };
-}
-
-
-#[macro_export]
-macro_rules! log_template_impl {
-    ($method:ident $logger:expr, $expr:expr) => {{
-        $logger.$method(|| $expr);
-    }};
-    ($method:ident $logger:expr, $expr:expr, $body:tt) => {{
-        let __logger = $logger.clone();
-        __logger.group_begin(|| $expr);
-        let out = $body;
-        __logger.group_end();
-        out
-    }};
-}
-
-#[macro_export]
-macro_rules! with_internal_bug_message { ($f:ident $($args:tt)*) => { $crate::$f! {
-"This is a bug. We will be thankful if you report it and provide us with as much information as \
-possible at https://github.com/luna/enso/issues. Thank you!"
-$($args)*
-}};}
-
-#[macro_export]
-macro_rules! log_internal_bug_template {
-    ($($toks:tt)*) => {
-        $crate::with_internal_bug_message! { log_internal_bug_template_impl $($toks)* }
-    };
-}
-
-#[macro_export]
-macro_rules! log_internal_bug_template_impl {
-    ($note:tt $method:ident $logger:expr, $message:tt $($rest:tt)*) => {
-        $crate::log_template_impl! {$method $logger,
-            format!("Internal Error. {}\n\n{}",iformat!($message),$note) $($rest)*
-        }
-    };
-}
-
-
-#[macro_export]
-macro_rules! info {
-    ($($toks:tt)*) => {
-        $crate::log_template! {info $($toks)*}
-    };
-}
-
-#[macro_export]
-macro_rules! warning {
-    ($($toks:tt)*) => {
-        $crate::log_template! {warning $($toks)*}
-    };
-}
-
-#[macro_export]
-macro_rules! internal_warning {
-    ($($toks:tt)*) => {
-        $crate::log_internal_bug_template! {warning $($toks)*}
-    };
-}
-
-
-
-
 // =============
 // === Utils ===
 // =============
